services/orchestrator_agent/requirements.txt
google-adk
python-dotenv
google-cloud-firestore
google-generativeai